Fix for sparse arrays
json_diff([], [,1]) did produce the following result:

  [{op:'add', path:'/0', value:undefined},
   {op:'add', path:'/1', value:1}]

This looks correct.  However, When converted to JSON, the undefined
value property would simply be ignored instead of turning into
'"value":null', so we have to explicitly cast undefined values to null.

The resulting patch will then produce the result [null,1], which in JSON
is equivalent to [,1].